Removed approval form modal and replaced with general form. Moved input modules into their own files. Added approval page to react router. Questions from database still populate on approval page.
