+++ src/browser/components/common/router.js
@@ -36,20 +36,23 @@ export default class Routes extends Component {
     }
 
     const LandingComponent = (props, state, params) =>
-    <Landing profile={fakeProfile} {...props}  stats={fakeStats} {...props}/>
+    <Landing profile={fakeProfile} {...props}  stats={fakeStats} />
 
     const GameOptionsComponent = (props, state, params) =>
-    <GameOptions profile={fakeProfile} {...props} stats={fakeStats} {...props} />
+    <GameOptions profile={fakeProfile} {...props} stats={fakeStats} />
 
     const ProfileComponent = (props, state, params) =>
-    <Profile profile={fakeProfile} {...props} stats={fakeStats} {...props} />
+    <Profile profile={fakeProfile} {...props} stats={fakeStats} />
+
+    const ApprovalComponent = (props, state, params) =>
+    <ApprovalPage profile={fakeProfile} stats={fakeStats} {...props} />
 
     return (
       <BrowserRouter>
         <div>
           <Route exact path='/' component={LandingComponent} />
           <Route path='/:topic/:difficulty' component={LandingComponent} />
-          <Route path='/approval' component={ApprovalPage} />
+          <Route path='/approval' component={ApprovalComponent} />
           <Route path='/profile' component={ProfileComponent} />
         </div>
       </BrowserRouter>
