+++ src/browser/components/pages/landing/index.js
@@ -1,12 +1,7 @@
 import React, {Component} from 'react'
-import QueryString from 'query-string'
-import {uniq, flatMap, take, shuffle} from 'lodash'
 
 import Layout from '../layout/index'
-import Game from '../game/index'
 import GameOptions from '../../molecules/game-options/index'
-import Header from '../../molecules/header/index'
-import Request from '../../common/requests'
 
 require('../../../../../public/stylesheets/uikit.min.css')
 
@@ -16,9 +11,7 @@ export default class Landing extends Component {
     this.state = {}
   }
 
-
   render() {
-
     return (
       <div className="uk-container">
         <Layout profile={this.props.profile} stats={this.props.stats}>
